% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackrock.R
\name{get_channel}
\alias{get_channel}
\title{Get channel data}
\usage{
get_channel(x, channel_id)
}
\arguments{
\item{x}{path \code{prefix} specified in \code{\link{import_nsp}}, or
\code{'nev/nsx'} object}

\item{channel_id}{integer channel number. Please be aware that channel
number, channel ID, electrode ID refer to the same concept in
'Blackrock' 'NEV' specifications. Electrodes are not physical metals, they
refer to channels for historical reasons.}
}
\value{
A list containing channel data and meta information, along with
the enclosing 'NSx' information; for invalid channel ID, this function
returns \code{NULL}
}
\description{
Obtain channel information and data from given prefix and
channel ID.
}
