% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackrock.R
\name{get_nev}
\alias{get_nev}
\title{Load 'NEV' information from path prefix}
\usage{
get_nev(x, ...)
}
\arguments{
\item{x}{path \code{prefix} specified in \code{\link{import_nsp}}, or
\code{'nev/nsx'} object}

\item{...}{reserved for future use}
}
\value{
'NEV' header information if \code{x} is valid, otherwise \code{NULL}.
See Section "'NEV' Data" in \code{\link{get_specification}}
}
\description{
Load 'NEV' information from path prefix
}
