% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackrock.R
\name{get_nsp}
\alias{get_nsp}
\title{Get a collection list containing 'NEV' and 'NSx' headers}
\usage{
get_nsp(x)
}
\arguments{
\item{x}{path \code{prefix} specified in \code{\link{import_nsp}}, or
\code{'nev/nsx'} object}
}
\value{
A list containing \code{'nev'} and imported \code{'nsx'} headers,
see \code{\link{import_nsp}} for details
}
\description{
Get a collection list containing 'NEV' and 'NSx' headers
}
