% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackrock.R
\name{get_nsx}
\alias{get_nsx}
\title{Load 'NSx' information from path prefix}
\usage{
get_nsx(x, which, ...)
}
\arguments{
\item{x}{path \code{prefix} specified in \code{\link{import_nsp}}, or
\code{'nev/nsx'} object}

\item{which}{which 'NSx' to load, for example, \code{which=3} loads
\code{'ns3'} headers}

\item{...}{reserved for future use}
}
\value{
'NSx' header information if data is found, otherwise returns
\code{NULL}. See Section "'NSx' Data" in \code{\link{get_specification}}
}
\description{
Load 'NSx' information from path prefix
}
