% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specifications.R
\name{get_specification}
\alias{get_specification}
\title{Get '.nev' or 'nsx' specification}
\usage{
get_specification(version, type = c("nev", "nsx"))
}
\arguments{
\item{version}{either character string or a vector of two integers; for
example, \code{"2.2"}, \code{"2.3"}, \code{c(3, 0)}. Currently only these
three versions are supported since I was unable to find any other versions.
Please file an issue ticket if others versions are desired.}

\item{type}{file type; choices are \code{'nev'} or \code{'nsx'}}
}
\value{
The file specification as a list. The specification usually contains
three sections: basic header (fixed length), extended header
(dictionary-style), and data packets (data stream). The specification is
used to parse the data files.
}
\description{
Get '.nev' or 'nsx' specification
}
\examples{

get_specification(version = c(2,3), type = "nev")

get_specification(version = "3.0", type = "nsx")

}
