% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackrock.R
\name{import_nsp}
\alias{import_nsp}
\title{Import signal data from 'Blackrock-Microsystems' data files}
\usage{
import_nsp(
  path,
  prefix = NULL,
  exclude_events = "spike",
  exclude_nsx = NULL,
  verbose = TRUE,
  partition_prefix = "/part"
)
}
\arguments{
\item{path}{path to 'NEV' or 'NSx' files}

\item{prefix}{path prefix to save data files into}

\item{exclude_events}{exclude one or more 'NEV' data events, choices are
\code{'spike'}, \code{'video_sync'}, \code{'digital_inputs'},
\code{'tracking'}, \code{'button_trigger'}, \code{'comment'},
\code{'configuration'}; default is \code{'spike'} since the spike data takes
very long time to parse, and most users might still use their own offline
spike-sorting algorithms.}

\item{exclude_nsx}{excluded 'NSx' types, integer vectors from 1 to 9; for
example, \code{exclude_nsx=c(1,2)} will prevent \code{'ns1'} and \code{'ns2'}
from being imported}

\item{verbose}{logical or a progress object: when logical, \code{verbose}
indicates whether to print the progress as messages;
when \code{verbose} is a progress object, this object must have \code{inc}
method; examples are \code{Progress} in the \code{shiny} package,
\code{progress2} from \code{dipsaus}, or \code{shiny_progress} from
\code{shidashi}}

\item{partition_prefix}{additional prefix to the data partition; default is
\code{"/part"}}
}
\value{
A list of configurations, see 'Details'.
}
\description{
Please use \code{import_nsp} to import 'NEV' and 'NSx' files.
}
