% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackrock.R
\name{import_nsp}
\alias{import_nsp}
\title{Import signal data from 'Blackrock-Microsystems' data files}
\usage{
import_nsp(
  path,
  prefix = NULL,
  exclude_events = "spike",
  exclude_nsx = NULL,
  verbose = TRUE,
  partition_prefix = "/part"
)
}
\arguments{
\item{path}{path to 'NEV' or 'NSx' files}

\item{prefix}{path prefix to save data files into}

\item{exclude_events}{exclude one or more 'NEV' data events, choices are
\code{'spike'}, \code{'video_sync'}, \code{'digital_inputs'},
\code{'tracking'}, \code{'button_trigger'}, \code{'comment'},
\code{'configuration'}; default is \code{'spike'} since the spike data takes
very long time to parse, and most users might still use their own offline
spike-sorting algorithms.}

\item{exclude_nsx}{excluded 'NSx' types, integer vectors from 1 to 9; for
example, \code{exclude_nsx=c(1,2)} will prevent \code{'ns1'} and \code{'ns2'}
from being imported}

\item{verbose}{logical or a progress object: when logical, \code{verbose}
indicates whether to print the progress as messages;
when \code{verbose} is a progress object, this object must have \code{inc}
method; examples are \code{Progress} in the \code{shiny} package,
\code{progress2} from \code{dipsaus}, or \code{shiny_progress} from
\code{shidashi}}

\item{partition_prefix}{additional prefix to the data partition; default is
\code{"/part"}}
}
\value{
A list of configurations, see 'Details'.
}
\description{
Please use \code{import_nsp} to import 'NEV' and 'NSx' files.
}
\section{'NEV' Data}{

A 'NEV' object consists of three sections:

Section 1 contains basic information such as the time-origin of all the
time-stamps, the time-stamp sampling frequency, data packets sizes.

Section 2 is extended header containing the configurations of channels,
digital signals, etc. For any data packets in section 3, there should
be at least one table in this section describing the settings.

section 3 is a collection of event packets such as digital signal inputs (
most likely to be used at version 2.2 or by 'Ripple'), spike waveform,
comments (sometimes storing epoch information), etc.

Please be aware that while most common entries can be found across different
file versions, some entries are version-specific. If you are making your
script general, you need to be very careful handling these differences.
For more information, please search for the data specification manual
from the 'Blackrock-Microsystems' website.
}

\section{'NSx' Data}{

A 'NSx' file refers to the data files ending with \code{'ns1'} through
\code{'ns9'}. Common types are \code{'ns2'} (sampling at 1000 Hz),
\code{'ns3'} (sampling at 2000 Hz), and \code{'ns5'} (sampling at 30,000 Hz).

A 'NSx' file also consists of three sections. Section 1 contains basic
information such as the time-origin of all the time-stamps, sampling
frequencies, and channel counts within the file. Please be careful that
item \code{time_resolution_timestamp} is not the sampling frequency for
signals. This item is the sampling frequency for time-stamp. To obtain the
signal sample rate, divided \code{time_resolution_timestamp} by
\code{period}. For example, \code{'ns3'} usually has time-stamp resolution
\code{30,000} and \code{period=15}, hence the signal sample rate is
\code{30000/15=2000Hz}.

Section 2 usually contains one and only one channel table of which the
number of rows should coincide with number of channels from section 1. Other
information such as channel labels, physical connectors, pins, units, filter
settings, digital-to-analog conversion are also included. Since
\code{readNSx} always attempts to convert signals in 'volts' or 'milli-volts'
to 'micro-volts', the \code{'units'} column might be different to what's
actual recorded in the 'NSx' file headers.

Section 3 contains partitions of continuous recording. When imported/loaded
from \code{readNSx}, the digital signals are always converted to analog
signals with 'micro-volts' unit. Please use \code{\link{get_channel}} to
get the channel data.
}

